Some Essentials
Setup a SystemManager which essentially defines instances of all the systems and stores them in a vector, makes it very easy to call update on all of them.
Also define a UserInputSystem which takes Event.Key.Code in update as well as the usual, need to integrate this into normal systems, so far it is called independent of rest of systems.
So have added a player, the controls add a velocity to just the player entity.