Asteroids spawn correctly
Asteroids now only spawn outside the window, but within the boundary. So spawning is never seen and asteroids just come in from the side of the screen.
This is achieved by first generating random position on the screen, then upscaling them to the bounding box-window box. Then adding the window box. Therefore mapping positions out of the window box but within the boundary box