Asteroids Collisions done
Asteroids-Asteroids collisions are now done.
When 2 asteroids hit, they are split into 2 halves of around half the size of the original. These 2 halves each go 30 degrees off from the resulting velocity of the collisions. These halves are spawned perpendicular from the direction of collisions from the original position (This simulates the asteroids splitting down the middle at the point of collision.