Changing back some stuff new design
There were too many circular dependencies, so instead of storing systems inside scene it will now go back to the system manager.
Scene Manager and Scene can reference system manager (this is the same way entities are added to registry). Then the scene can store the IDs of the systems it adds.